validation update

+++ src/App.jsx
@@ -8,11 +8,36 @@ import Recipe from './component/Recipe/Recipe'
 import RecipeItem from './component/Recipe/RecipeItem'
 
 function App() {
+
   const [items, setItem] = useState([])
+
   const handleCart = (item) => {
-    const newItem = [...items, item]
-    setItem(newItem)
+    const isExist = items.find(cart => cart.recipe_id == item.recipe_id)
+    if (!isExist) {
+      const newItem = [...items, item]
+      setItem(newItem)
+    } else {
+      alert('jhgdfkls')
+    }
+
   }
+
+
+
+
+
+
+
+  const handleCook = (recipeId) => {
+
+    const newRecipe = items.filter(recipe => recipe.recipe_id != recipeId)
+    setItem(newRecipe)
+
+
+  }
+
+
+
 
   return (
     <>
@@ -21,7 +46,7 @@ function App() {
       <Recipe></Recipe>
       <div className='flex w-[81%] mx-auto my-10 gap-3'>
         <RecipeItem handleCart={handleCart}></RecipeItem>
-        <Cook items={items}></Cook>
+        <Cook handleCook={handleCook} items={items}></Cook>
       </div>
 
     </>
